Add a menu for switching between release transl*ations
This info is sourced from the transl-tracklisting relationship type:
http://musicbrainz.org/relationship/fc399d47-23a7-4c28-bfcf-0607a562b644

Not every transl* release will be a pseudo-release, but the ones that are
are now excluded from the "Other versions" menu.

If a pseudo-release is loaded directly into Picard, Picard will try to load
one of the linked source releases instead, but use the artists and titles from
the pseudo-release. The new 'musicbrainz_translreleaseid' tag points to the
pseudo-release.